gen-ai-course/14_ai_projects/smart_notion_sync_agent/pyproject.toml
[project]
name = "smart-notion-sync-agent"
version = "0.1.0"
description = "Bi-directional sync between Google Workspace and Notion with LLM conflict resolution"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "langchain-groq>=0.1.0",
    "langchain-core>=0.2.0",
    "langgraph>=0.0.40",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "notion-client>=2.2.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "slack_sdk>=3.30.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/notion_sync"]

[tool.hatch.build.targets.sdist]
include = ["/src"]
